Lambda sink refactor
