+++ data-prepper-plugins/aws-lambda/src/test/resources/simplelogger.properties
