+++ settings.gradle
@@ -178,4 +178,4 @@ include 'data-prepper-plugins:mongodb'
 include 'data-prepper-plugins:rds-source'
 include 'data-prepper-plugins:http-source-common'
 include 'data-prepper-plugins:http-common'
-include 'data-prepper-plugins:lambda'
+include 'data-prepper-plugins:aws-lambda'
